refactor(navbar): move navbar to storybook-static folder
re #51
